Bug fixed in error computation

+++ PMF.py
@@ -89,8 +89,8 @@ class PMF:
 
                 # Compute Objective Function after
                 if batch == self.num_batches - 1:
-                    pred_out = np.sum(np.multiply(self.w_I[np.array(train_vec[:,0], dtype='int8'),:],
-                                                    self.w_C[np.array(train_vec[:,1], dtype='int8'),:]),
+                    pred_out = np.sum(np.multiply(self.w_I[np.array(train_vec[:,0], dtype='int32'),:],
+                                                    self.w_C[np.array(train_vec[:,1], dtype='int32'),:]),
                                         axis=1) # mean_inv subtracted
                     rawErr = pred_out - train_vec[:, 2] + self.mean_inv
                     obj = LA.norm(rawErr) ** 2 \
@@ -100,8 +100,8 @@ class PMF:
 
                 # Compute validation error
                 if batch == self.num_batches - 1:
-                    pred_out = np.sum(np.multiply(self.w_I[np.array(val_vec[:,0], dtype='int8'),:],
-                                                    self.w_C[np.array(val_vec[:,1], dtype='int8'),:]),
+                    pred_out = np.sum(np.multiply(self.w_I[np.array(val_vec[:,0], dtype='int32'),:],
+                                                    self.w_C[np.array(val_vec[:,1], dtype='int32'),:]),
                                         axis=1) # mean_inv subtracted
                     rawErr = pred_out - val_vec[:, 2] + self.mean_inv
                     self.err_val.append(LA.norm(rawErr)/np.sqrt(pairs_va))
